requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
matplotlib>=3.0.0
pandas>=2.0.0
pyodbc>=5.0.0

# Core LangChain 2026 Ecosystem
langchain>=1.2.0
langchain-openai>=1.1.7
langchain-community>=0.4.0
langgraph>=1.0.6
langchain-core>=1.2.7

# Fixed Azure Version (Stable)
azure-search-documents>=11.6.0
azure-core>=1.31.0

# Other
ddgs>=0.0.1
